Updating 2nd autoupgrade model

+++ application/models/upgrade2_model.php
@@ -41,6 +41,9 @@ class Upgrade2_model extends CI_Model {
 		if (!is_writable('content/themes/default')) {
 			return FALSE;
 		}
+		if (!is_writable('content/themes/fuuka')) {
+			return FALSE;
+		}
 		if (!is_writable('content/cache')) {
 			return FALSE;
 		}
@@ -88,6 +91,9 @@ class Upgrade2_model extends CI_Model {
 		if (!file_exists('content/cache/upgrade/content/themes/default')) {
 			return FALSE;
 		}
+		if (!file_exists('content/cache/upgrade/content/themes/fuuka')) {
+			return FALSE;
+		}
 
 		unlink('index.php');
 		rename('content/cache/upgrade/index.php', 'index.php');
@@ -99,6 +105,8 @@ class Upgrade2_model extends CI_Model {
 		rename('content/cache/upgrade/assets', 'assets');
 		delete_files('content/themes/default/', TRUE);
 		rename('content/cache/upgrade/content/themes/default', 'content/themes/default');
+		delete_files('content/themes/fuuka/', TRUE);
+		rename('content/cache/upgrade/content/themes/fuuka', 'content/themes/fuuka');
 		
 		// delete the mobile folder for a while... 05/09/2011
 		//if (file_exists('content/themes/mobile')) {
